chore: rename use-service-router folder
