+++ apps/web/src/services/asset-inventory/components/CloudServiceDetailMultipleSelectedData.vue
@@ -28,7 +28,7 @@ import type { CloudServiceListParameters } from '@/schema/inventory/cloud-servic
 import type { CloudServiceModel } from '@/schema/inventory/cloud-service/model';
 import { i18n } from '@/translations';
 
-import { useServiceRouter } from '@/router/helpers/use-service-router';
+import { useServiceRouter } from '@/router/composables/use-service-router';
 
 import { useAppContextStore } from '@/store/app-context/app-context-store';
 import { useUserWorkspaceStore } from '@/store/app-context/workspace/user-workspace-store';
