+++ apps/web/src/services/asset-inventory/components/CloudServiceDetailTabs.vue
@@ -14,7 +14,7 @@ import type { CloudServiceGetParameters } from '@/schema/inventory/cloud-service
 import type { CloudServiceModel } from '@/schema/inventory/cloud-service/model';
 import { i18n } from '@/translations';
 
-import { useServiceRouter } from '@/router/helpers/use-service-router';
+import { useServiceRouter } from '@/router/composables/use-service-router';
 
 import { useGlobalConfigSchemaStore } from '@/store/global-config-schema/global-config-schema-store';
 import { useAllReferenceStore } from '@/store/reference/all-reference-store';
